Milestone 4 (part 4): transactions
Solves the two problems identified when transactions were first
investigated and deliberately deferred:

1. Scope mismatch: SQLite calls a vtab module's xBegin once per *(vtab
   instance, transaction)* - three vgi_worker tables from the same
   (location, catalog) in one SQL transaction get three xBegin calls -
   but VGI's catalog_transaction_begin is scoped to one *(location,
   catalog) attachment*. Solved with the same shared, ref-counted-per-key
   pattern ConnectionPool already uses for attach_opaque_data: new
   Begin/Commit/RollbackTransaction only call the real RPC on the first
   Begin/last Commit-or-Rollback for a key; CurrentTransactionOpaqueData(key)
   lets every bind made during the transaction (TableScanner::Bind,
   TableWriter::Write - both gained an optional transaction_opaque_data
   parameter) thread it through regardless of which physical connection
   that particular checkout happens to be, matching the same
   reuse-across-connections precedent attach_opaque_data already
   established. Gated on CatalogAttachResult.supports_transactions
   (cached in ConnectionPool's new attach_info_ map alongside
   attach_opaque_data, populated by the xConnect that always precedes any
   possible xBegin on a table) - a catalog that doesn't support
   transactions never gets a catalog_transaction_begin call at all.

2. Protocol shape mismatch: VGI has no nested-savepoint RPCs at all, only
   flat begin/commit/rollback. Resolved by simply not implementing
   SQLite's xSavepoint/xRelease/xRollbackTo (left null) - confirmed safe
   rather than assumed: every INSERT/UPDATE/DELETE test had already been
   passing with these three null the whole time before transactions
   existed, proving SQLite doesn't require a vtab to implement them for
   ordinary transactions; only explicit SQL SAVEPOINT nesting would need
   them.

vgi_vtab.cpp gained xBegin/xSync (a no-op - VGI has no two-phase pre-commit
step)/xCommit/xRollback as thin ConnectionPool passthroughs. A real,
previously-unknown fact surfaced immediately on registering them at all:
xBegin fires for *every* vgi_worker table touched inside a transaction,
not just written ones - including a plain SELECT inside SQLite's own
implicit per-statement transaction. The `example` catalog (used by nearly
every read test in this suite) declares supports_transactions=True, so
every pre-existing read test immediately started exercising the new
coordination code path too - and all of them kept passing.

test/integration/test_transactions.py adds explicit-transaction coverage:
commit, rollback, sequential transactions on one catalog, and - the
scenario the ref-counted coordination specifically exists for - one
explicit transaction spanning two vgi_worker tables from the same
catalog, asserting on the actual data read from both rather than just "no
exception was raised" (a wrong implementation could silently open two
unrelated transactions without erroring). 39/39 integration tests passing
(33 pre-existing + 6 new).

Built and verified on a fresh Linux/aarch64 environment (Amazon Linux,
GCC 11) for the first time, at the user's direction, not just macOS/clang
- this surfaced several real, independent portability bugs never
exercised by a from-scratch build before, all fixed rather than worked
around locally:
- arrow::MakeScalar(sqlite3_value_int64(...)) failed template argument
  deduction on GCC: sqlite3_int64 (effectively `long long`) and int64_t
  (`long` on this LP64 Linux target) are distinct types to GCC's stricter
  deduction even though both are 64-bit - clang accepted the ambiguity,
  GCC didn't. Fixed with explicit static_cast<int64_t>(...) at every call
  site across vgi_vtab.cpp and type_mapping.cpp.
- arrow::RecordBatch::Make(..., {single_array}) in vgi_vtab.cpp's DoDelete
  failed to compile alongside the above - turned out to be a cascading
  failure from the same broken expression, not an independent bug (every
  other brace-init RecordBatch::Make call site elsewhere in this repo
  compiled fine on GCC once the real MakeScalar bug was fixed).
- vgi_sqlite_driver (and, transitively, vgi-c++'s and vgi-rpc-c++'s own
  static libraries) had never been built position-independent - macOS's
  linker tolerates a non-PIC static library linked into a shared module,
  Linux's refuses outright ("recompile with -fPIC"). Fixed by setting
  CMAKE_POSITION_INDEPENDENT_CODE ON globally in this repo's own
  CMakeLists.txt, before pulling in vgi-c++/vgi-rpc-c++ via
  add_subdirectory - applies to every target built in that subtree without
  needing to touch either sibling repo's own CMakeLists.txt.

+++ src/types/type_mapping.cpp
@@ -344,7 +344,7 @@ std::shared_ptr<arrow::Scalar> BuildArrowScalarFromSqliteValue(
         case arrow::Type::UINT32:
         case arrow::Type::UINT64: {
             if (sqlite3_value_type(value) != SQLITE_INTEGER) return nullptr;
-            auto scalar = arrow::MakeScalar(sqlite3_value_int64(value));
+            auto scalar = arrow::MakeScalar(static_cast<int64_t>(sqlite3_value_int64(value)));
             auto cast_result = scalar->CastTo(target_type);
             return cast_result.ok() ? cast_result.ValueUnsafe() : nullptr;
         }
